fix-prettier & eslint
