+++ frontend/src/components/Board/BoardTitleSection/DotsButton/DotsButton.jsx
@@ -1,10 +1,10 @@
-import {useState} from 'react';
+import { useState } from 'react';
 import { StyledDotsButton, ButtonContainer } from './DotsButton.styled';
 import { FaEllipsisH } from 'react-icons/fa';
 
 import PropTypes from 'prop-types';
 
-const DotsButton = ({children}) => {
+const DotsButton = ({ children }) => {
   const [isOpen, setIsOpen] = useState(false);
 
   return (
@@ -14,11 +14,11 @@ const DotsButton = ({children}) => {
       </ButtonContainer>
       {isOpen && children}
     </StyledDotsButton>
-  )
-}
+  );
+};
 
 DotsButton.propTypes = {
-  children: PropTypes.node
-}
+  children: PropTypes.node,
+};
 
 export default DotsButton;
