+++ frontend/src/components/Board/BoardTitleSection/DotsButton/DotsButton.styled.jsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import {breakpoints} from "components/../rwd/breakpoints";
+import { breakpoints } from 'rwd/breakpoints';
 
 export const StyledDotsButton = styled.div`
   position: absolute;
@@ -11,25 +11,25 @@ export const StyledDotsButton = styled.div`
   align-items: center;
   color: #bbb;
   cursor: pointer;
-`
+`;
 
 export const ButtonContainer = styled.div`
   color: #bbb;
   transition: color 0.15s ease-out;
-  
-    &:hover {
-      color: #6b6b6b;
-    }
-  
-    @media ${breakpoints.md} {
-      font-size: 30px;
-    }
-  
-    @media ${breakpoints.lg} {
-      font-size: 40px;
-    }
-  
-    @media ${breakpoints.xl} {
-      font-size: 50px;
-    }
-`
+
+  &:hover {
+    color: #6b6b6b;
+  }
+
+  @media ${breakpoints.md} {
+    font-size: 30px;
+  }
+
+  @media ${breakpoints.lg} {
+    font-size: 40px;
+  }
+
+  @media ${breakpoints.xl} {
+    font-size: 50px;
+  }
+`;
